+++ src/commands/index.ts
@@ -2,7 +2,7 @@ export * from "./command-fablab";
 export * from "./create-embed-message";
 export * from "./help";
 export * from "./logs-parameters";
-export * from "./manage-channel";
+export * from "./manage-webhook";
 export * from "./process-infos";
 export * from "./role-reaction";
 export * from "./site-fablab";